chore: move utility function in integration into utils
